pyproject.toml
[project]
name = "ctd_pre_system"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Magnus Wenzer", email = "nodc@smhi.se"},
]
dependencies = [
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "psutil>=5.9.8",
    "svepa @ git+https://github.com/sharksmhi/svepa.git@v1.0.0",
    "file-explorer @ git+https://github.com/sharksmhi/file_explorer.git",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["."]
